Don't exit when RemoteMemoryRead fails
This was causing QEMU to exit prematurely during replay. The reason (we think) is that the CPU thread was able to run for a short time between the end of the replay and QEMU's graceful shutdown. During that time, the CPU thread would try to access MMIO, and would hit the RemoteMemoryRead failure, and immediately kill QEMU. This caused problems because it prevented other QEMU shutdown code from being called, in particular the code that flushes out the pandalog.